storage: sqlite: add "_parser" column to "event" table
The "parser" (default) field of each event is currently store serialized
in the SQLite storage file. That is, it is not possible to quickly extract
all events, with psort, for a given parser. The only way is to write a
filter expression like 'parser is "filestat"'. The latter requires to
process every single events in the storage. This is inefficient and it does
not benefit from the SQLite backend facilities.

This patch adds an extra "_parser" column to the "event" table to allow
searching through the SQLite storage file with "WHERE _parser='filestat'"
requests or alike.

Note that it breaks compatibility both in reading and writing with previous
SQLite storage file.