storage: introduce "parser" argument to GetSortedEvents() methods
This patch adds the "parser" argument to all GetSortedEvents() methods and
provides a concrete implementation for SQLite file storage. Events can
still be filtered using the time range and are ordered by timestamp.

The only "logic" part is in the "sqlite_file.py" which add a new SQL
query (note the '/' in the LIKE clause):

    (_parser = "{0:s}" OR _parser LIKE "%/{1:s}")

The query both works for exact match (e.g. "appcompatcache") and parser
name prefixed by a plugin name (e.g. "winreg/appcompatcache"). The
rationale behind this is that "pinfo.py" currently list parsers by
omitting the plugin name and can be used as a basis to generate filter
expressions.

Furthermore, the EventObject "parser" attribute contains the plugin prefix
(if any), so does the "_parser" column in the SQLite storage file. Note
that some parsers, like "filestat" does not have any plugin prefix.

The Redis interfaces have been updated as well but are currently not
supported (just like the "time_range" argument).